requirements.txt
django
Djangorestframework
psycopg2
drf_spectacular
pillow